# Use indirect relative func-ptr for harvard archs