--- a/include/swift/ABI/Executor.h
+++ b/include/swift/ABI/Executor.h
@@ -241,7 +241,7 @@ template <class AsyncSignature>
 class AsyncFunctionPointer {
 public:
   /// The function to run.
-  RelativeDirectPointer<AsyncFunctionType<AsyncSignature>,
+  RelativeFunctionPointer<AsyncFunctionType<AsyncSignature>,
                         /*nullable*/ false,
                         int32_t> Function;