--- a/lib/IRGen/ConstantBuilder.h
+++ b/lib/IRGen/ConstantBuilder.h
@@ -91,11 +91,6 @@ public:
 
   void addRelativeAddress(llvm::Constant *target) {
     assert(!isa<llvm::ConstantPointerNull>(target));
-    if (IGM().TargetInfo.OutputObjectFormat == llvm::Triple::Wasm) {
-      // WebAssembly: hack: doesn't support PCrel data relocations
-      add(llvm::ConstantExpr::getPtrToInt(target, IGM().RelativeAddressTy, false));
-      return;
-    }
     addRelativeOffset(IGM().RelativeAddressTy, target);
   }
 
@@ -104,21 +99,6 @@ public:
   /// a "GOT-equivalent", i.e. a pointer to an external object; if so,
   /// set the low bit of the offset to indicate that this is true.
   void addRelativeAddress(ConstantReference reference) {
-    if (IGM().TargetInfo.OutputObjectFormat == llvm::Triple::Wasm) {
-      // WebAssembly: hack: doesn't support PCrel data relocations
-      // also, we should set the lowest bit, but I don't know how to do that
-      // there's no GOT on WebAssembly anyways though
-
-
-      llvm::Constant *offset = llvm::ConstantExpr::getPtrToInt(reference.getValue(), IGM().RelativeAddressTy, false);
-      // borrowed from addTaggedRelativeOffset
-      unsigned tag = unsigned(reference.isIndirect());
-      if (tag) {
-        offset = llvm::ConstantExpr::getAdd(offset, llvm::ConstantInt::get(IGM().RelativeAddressTy, tag));
-      }
-      add(offset);
-      return;
-    }
     addTaggedRelativeOffset(IGM().RelativeAddressTy,
                             reference.getValue(),
                             unsigned(reference.isIndirect()));
@@ -128,11 +108,6 @@ public:
   /// The target must be a "GOT-equivalent", i.e. a pointer to an
   /// external object.
   void addIndirectRelativeAddress(ConstantReference reference) {
-    if (IGM().TargetInfo.OutputObjectFormat == llvm::Triple::Wasm) {
-      // WebAssembly: hack: doesn't support PCrel data relocations
-      add(llvm::ConstantExpr::getPtrToInt(reference.getValue(), IGM().RelativeAddressTy, false));
-      return;
-    }
     assert(reference.isIndirect());
     addRelativeOffset(IGM().RelativeAddressTy,
                       reference.getValue());