--- a/lib/IRGen/IRGenFunction.cpp
+++ b/lib/IRGen/IRGenFunction.cpp
@@ -330,13 +330,6 @@ IRGenFunction::emitLoadOfRelativePointer(Address addr, bool isFar,
     value = Builder.CreateSExt(value, IGM.IntPtrTy);
   }
 
-  // FIXME: add absolute pointer mode for IRGen
-  if (IGM.TargetInfo.OutputObjectFormat == llvm::Triple::Wasm) {
-    auto *uncastPointer = Builder.CreateIntToPtr(value, IGM.Int8PtrTy);
-    auto uncastPointerAddress = Address(uncastPointer, IGM.getPointerAlignment());
-    auto pointer = Builder.CreateBitCast(uncastPointerAddress, expectedType);
-    return pointer.getAddress();
-  }
   auto *addrInt = Builder.CreatePtrToInt(addr.getAddress(), IGM.IntPtrTy);
   auto *uncastPointerInt = Builder.CreateAdd(addrInt, value);
   auto *uncastPointer = Builder.CreateIntToPtr(uncastPointerInt, IGM.Int8PtrTy);