--- a/lib/IRGen/IRGenFunction.h
+++ b/lib/IRGen/IRGenFunction.h
@@ -266,6 +266,10 @@ public:
   llvm::Value *emitLoadOfRelativePointer(Address addr, bool isFar,
                                          llvm::PointerType *expectedType,
                                          const llvm::Twine &name = "");
+  llvm::Value *
+  emitLoadOfRelativeFunctionPointer(Address addr, bool isFar,
+                                    llvm::PointerType *expectedType,
+                                    const llvm::Twine &name = "");
 
   llvm::Value *emitAllocObjectCall(llvm::Value *metadata, llvm::Value *size,
                                    llvm::Value *alignMask,