… cmake/caches/Runtime-WASI-wasm32.cmake
… lib/IRGen/MetadataRequest.cpp
… stdlib/cmake/modules/StdlibOptions.cmake
… stdlib/public/core/KeyPath.swift
… stdlib/public/runtime/Metadata.cpp
… stdlib/public/runtime/MetadataLookup.cpp