I measured the runtime performance and binary size benchmark against [my toy linker](https://github.com/kateinoigakukun/chibi-link) with quite large object inputs

It seems there is no critical regression in runtime performance.

```console
chibi-link ((ddb57c0...) $)$ hyperfine "wasmtime run --mapdir /::./objects .build-new/wasm32-unknown-wasi/debug/chibi-link.wasm -- main.o lib.o library.so.wasm -o main.wasm" "wasmtime run --mapdir /::./objects .build-old/wasm32-unknown-wasi/debug/chibi-link.wasm -- main.o lib.o library.so.wasm -o main.wasm"  --cleanup "rm ./objects/main.wasm" --max-runs 4
Benchmark 1: wasmtime run --mapdir /::./objects .build-new/wasm32-unknown-wasi/debug/chibi-link.wasm -- main.o lib.o library.so.wasm -o main.wasm
  Time (mean ± σ):      4.250 s ±  0.032 s    [User: 4.139 s, System: 0.052 s]
  Range (min … max):    4.216 s …  4.292 s    4 runs

Benchmark 2: wasmtime run --mapdir /::./objects .build-old/wasm32-unknown-wasi/debug/chibi-link.wasm -- main.o lib.o library.so.wasm -o main.wasm
  Time (mean ± σ):      4.251 s ±  0.021 s    [User: 4.182 s, System: 0.054 s]
  Range (min … max):    4.233 s …  4.279 s    4 runs

Summary
  'wasmtime run --mapdir /::./objects .build-new/wasm32-unknown-wasi/debug/chibi-link.wasm -- main.o lib.o library.so.wasm -o main.wasm' ran
    1.00 ± 0.01 times faster than 'wasmtime run --mapdir /::./objects .build-old/wasm32-unknown-wasi/debug/chibi-link.wasm -- main.o lib.o library.so.wasm -o main.wasm'

```

Of course, it adds size overheads for indirect trampoline pointers, but it's less than 1%.

```
# before wasm-strip
$ ls -l .build-*/wasm32-unknown-wasi/debug/*.wasm
-rwxrwxr-x 1 katei katei 7990379 Mar 24 05:38 .build-old/wasm32-unknown-wasi/debug/chibi-link.wasm
-rwxrwxr-x 1 katei katei 8009400 Mar 24 05:39 .build-new/wasm32-unknown-wasi/debug/chibi-link.wasm

# after wasm-strip
$ ls -l .build-*/wasm32-unknown-wasi/debug/*.wasm
-rwxrwxr-x 1 katei katei 5997860 Mar 24 05:57 .build-old/wasm32-unknown-wasi/debug/chibi-link.wasm
-rwxrwxr-x 1 katei katei 6016784 Mar 24 05:57 .build-new/wasm32-unknown-wasi/debug/chibi-link.wasm
```

My only concern here is that this might make VFE (virtual-function-elimination) less effective due to additional indirection, but I think we can improve constant expression evaluator to track indirect relative pointer if it's not supported.